macOS Seatbelt strict profile + MCP argv-first + watchdog cast 修复
- macOS: 重写 Seatbelt 策略为 deny-default + 精确 allowlist
  - 移除全局 (allow file-read*)，改为系统最小路径 + fs_readonly/fs_readwrite
  - fs_readonly 现在生效（之前完全未使用）
  - allow_fork=false 时不再生成 (allow process-fork)
  - Mach 服务精确白名单（6 个必需服务）
  - 路径规范化处理 macOS 符号链接
- MCP: execute_command 新增 argv 字段，优先使用 argv 传参
  - 保留 command 字段向后兼容
  - argv 通过 SDK exec() 直接传递，无 shell 解析
- watchdog: proc_pidrusage 参数类型修复

+++ crates/mimobox-os/src/macos.rs
@@ -223,7 +223,7 @@ fn spawn_memory_watchdog(
             let ret = unsafe {
                 proc_pidrusage(
                     pid,
-                    RUSAGE_INFO_V2 as i32,
+                    RUSAGE_INFO_V2,
                     &mut rui as *mut _ as *mut libc::rusage_info_t,
                 )
             };
